--- js/commentSentences.js
@@ -1,29 +1,0 @@
-// Список возможных предложений для комментариев
-
-export const commentSentences = [
-  'Отличная фотография!',
-  'Классный ракурс!',
-  'Красивая картинка!',
-  'Прекрасное место!',
-  'Хорошо передана атмосфера!',
-  'Очень интересно!',
-  'Очаровательная фотография!',
-  'Какое мастерство!',
-  'Замечательный снимок!',
-  'Очень красиво!',
-  'Удачный момент!',
-  'Супер!',
-  'Как же красиво!',
-  'Великолепно!',
-  'Приятная композиция!',
-  'Супер ракурс!',
-  'Мне нравится!',
-  'Прекрасное фото!',
-  'Хорошее качество!',
-  'Красивые цвета!',
-  'Завораживает!',
-  'Какая атмосфера!',
-  'Отличный фокус!',
-  'Чудесная работа!',
-  'Сильно!'
-];
