+++ js/main.js
@@ -1,2 +1,5 @@
 
 import './util.js';
+import {generateComments} from './data.js';
+
+generateComments(2);
